+++ recipes/bootstrap_github_action.php
@@ -174,7 +174,7 @@ task(
     'deploy:setup-permissions',
     function () {
         // folders
-        run("cd {{release_path}} && find . -type d -not \( -path './awstats*' -o -path './.git*' \) ! -perm 0755 -exec chmod 755 {} \;");
+        run("cd {{release_path}} && find . -type d -not \( -path './awstats*' -o -path './.git*' -o -path \"./var/cache*\" \) ! -perm 0755 -exec chmod 755 {} \;");
 
         // files
         run("cd {{release_path}} && find . -type f -not \( -path './awstats/*' -o -path './logs/*' -o -path './stats/*' -o -path './.git/*' \) ! -perm 0644 -exec chmod 644 {} \;");